pyproject.toml
[project]
name = "expense-tracker-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.1",
    "fastmcp>=2.14.5",
    "mysql-connector-python>=9.5.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]
